Improvements to use a config file instead of replacing
